+++ src/bot/helper.ts
@@ -1447,6 +1447,7 @@ function createBotHelper<
                 if (exchangeData.status === StatusEnum.notok) {
                   this.handleWarn(`Cannot get order ${exchangeData.reason}`)
                 } else {
+                  this.clearOrderStrikes(o)
                   const updatedOrder = await this.mergeCommonOrderWithOrder(
                     exchangeData.data,
                     o,